// src/enums/delivery-status.enum.ts
export enum DeliveryStatus {
  ALL = 'all',
  PENDING = 'pending',
  CONFIRMED = 'confirmed',
  ASSIGNED = 'assigned',
  PICKED_UP = 'picked_up',
  IN_TRANSIT = 'in_transit',
  ARRIVING = 'arriving',
  OUT_FOR_DELIVERY = 'out_for_delivery',
  DELIVERED = 'delivered',
  COMPLETED = 'completed',
  FAILED = 'failed',
  RETURNED = 'returned',
  CANCELLED = 'cancelled',
  VOIDED = 'voided'
} 